requirements.txt
asgiref==3.9.1
certifi==2025.10.5
dj-database-url==3.0.1
Django==5.2.6
gunicorn==23.0.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.1
sentry-sdk==2.40.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
sentry-sdk[django]==1.45.0
pytest==8.4.1
pytest-django==4.9.0
pytest-mock==3.14.0
